fzf install: exit if already installed, prevents error

+++ fzf/install.sh
@@ -3,6 +3,8 @@
 
 FZF_TMP="/tmp/fzf-install-src"
 
+if [[ -f "$HOME/.local/bin/fzf" ]]; then exit 0; fi
+
 set -e
 rm -rf "$FZF_TMP" || true
 git clone --depth 1 https://github.com/junegunn/fzf.git "$FZF_TMP"
@@ -11,7 +13,7 @@ git clone --depth 1 https://github.com/junegunn/fzf.git "$FZF_TMP"
 yes | "$FZF_TMP/install"
 
 mkdir -p "$HOME/.local/bin"
-cp -f "$FZF_TMP/bin/fzf" "$HOME/.local/bin"
+cp -f "$FZF_TMP/bin/fzf" "$HOME/.local/bin/fzf"
 
 rm -rf "$FZF_TMP"
